requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.0
cv==1.0.0
cvzone==1.6.1
cycler==0.12.1
Django==5.0
djangorestframework==3.14.0
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.12.2
idna==3.6
Jinja2==3.1.2
kiwisolver==1.4.5
kombu==5.3.4
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
opencv-python==4.9.0.80
packaging==23.2
pandas==2.1.4
pillow==10.2.0
prompt-toolkit==3.0.43
psutil==5.9.7
py-cpuinfo==9.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
scipy==1.11.4
seaborn==0.13.1
six==1.16.0
sqlparse==0.4.4
sympy==1.12
thop==0.1.1.post2209072238
torch==2.1.2+cpu
torchvision==0.16.2+cpu
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.0.232
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
